upgraded projects, got rid of some warnings and added a few toolbar icons I needed

+++ UIKit/Classes/UISegmentedControl.m
@@ -380,7 +380,7 @@ static NSString *kSSSegmentedControlEnabledKey = @"enabled";
         return nil;
     }
     
-    NSString *key = [NSString stringWithFormat:@"%lu", index];
+    NSString *key = [NSString stringWithFormat:@"%lu", (unsigned long)index];
     return [_segmentMeta objectForKey:key];
 }
 
@@ -405,7 +405,7 @@ static NSString *kSSSegmentedControlEnabledKey = @"enabled";
         _segmentMeta = [[NSMutableDictionary alloc] init];
     }
     
-    [_segmentMeta setValue:meta forKey:[NSString stringWithFormat:@"%lu", index]];
+    [_segmentMeta setValue:meta forKey:[NSString stringWithFormat:@"%lu", (unsigned long)index]];
     [self setNeedsDisplay];
 }
